Fix some untested matrix cases and add new matrix utilities

+++ tools/buildwheel.py
@@ -11,6 +11,13 @@ urls = [
     "https://files.pythonhosted.org/packages/9a/9a/36f71797fbaf1f4b6cb8debe1ab6d3ec969e8dbc651181f131a16b794d80/pymdown_extensions-9.0-py3-none-any.whl",  # noqa: E501
     "https://files.pythonhosted.org/packages/9f/52/801913c6da8bb8f3657e8f15dafb450079f225d8564ada518bb545812417/coloraide_extras-0.2.0-py3-none-any.whl"  # noqa: E501
 ]
+
+# Keep a compatible version in the repository so we can always release docs.
+# If we ever have a breaking change, we can patch `coloraide-extras` before
+# the release, locally upload it and then make a release.
+# Afterwards, once `coloraide` is available, we can release the official
+# `coloraide-extras`. This prevents our documentation from ever breaking.
+keep = {'coloraide_extras-0.2.0-py3-none-any.whl'}
 output = 'docs/src/markdown/playground/'
 externals = {}
 for url in urls:
@@ -57,7 +64,7 @@ def download_wheel(url, dest):
 if __name__ == "__main__":
     # Clean up all old wheels
     for file in glob.glob(output + '*.whl'):
-        if file not in externals.keys():
+        if file not in externals.keys() and os.path.basename(file) not in keep:
             os.remove(file)
 
     # Clean up build directory
@@ -69,6 +76,9 @@ if __name__ == "__main__":
         sys.exit(1)
 
     for file, url in externals.items():
+        if os.path.basename(file) in keep and os.path.exists(file):
+            print('Skipping: {}'.format(file))
+            continue
         if download_wheel(url, file):
             sys.exit(1)
